passing tests and renamed sign up to register
