+++ client/src/App.js
@@ -12,8 +12,8 @@ import { Routes, Route, Navigate } from "react-router-dom";
 //components
 
 import Login from "./components/login-and-register/Login";
-import Register from "./components/login-and-register/signUp";
-import RegisterConfirmation from "./components/login-and-register/signUpConfirmation";
+import Register from "./components/login-and-register/Register";
+import RegisterConfirmation from "./components/login-and-register/RegisterConfirmation";
 import Navbar from "./components/navbar/Navbar";
 import Home from "./components/home/Home";
 import Dashboard from "./components/Dashboard";
