+++ client/src/components/login-and-register/Login.js
@@ -28,7 +28,7 @@ function Login({ setAuth }) {
       });
 
       const parseRes = await response.json();
-      console.log(parseRes);
+      console.log('Login file line 31 => ', parseRes);
       if (parseRes.jwtToken) {
         localStorage.setItem("token", parseRes.jwtToken);
