+++ client/src/components/login-and-register/RegisterConfirmation.js
