+++ package.json
@@ -33,6 +33,7 @@
   },
   "devDependencies": {
     "cypress": "^12.3.0",
+    "msw": "^1.2.1",
     "nodemon": "^2.0.20"
   }
 }
